Inertia (#4)
- Inertia setup with Larave/Vue, all dockerized



* php+composer in docker

* installed inertia

* laravel new project

* env

* base project

* wip

* inertia setup

* setup worksgit add .

* setup

* more progress

* ça marche enfin pour vrai

* fix

* fix github actions

* prettier

* fix precommit

* fixes

* eslint vs prettier

* fix

* clean

* cleaner setup

* prettier missing plugin + readme

* fix

* fix

+++ docker/setup.sh
@@ -1,7 +1,22 @@
 #!/bin/bash
-cp docker/dev/.env.dev .env
-echo "DEV_USER_ID="$(id -u) >> .env
-echo "DEV_GROUP_ID="$(id -g) >> .env
-docker compose up -d --build --remove-orphans --quiet-pull 
-make npm-install
+if [ ! -f ".env" ]; then     
+    echo ✏️ initiating .env
+    cp docker/dev/.env.dev .env
+    echo "DEV_USER_ID="$(id -u) >> .env
+    echo "DEV_GROUP_ID="$(id -g) >> .env
+else 
+    echo ⚠️ .env already exists. Skipping
+fi
+source .env
+docker compose build 
+docker compose up -d --remove-orphans --quiet-pull
+#docker compose exec -T php composer install
+#docker compose exec -T php php artisan key:generate
+#docker compose exec -T php php artisan migrate --force
+#docker compose exec -T php npm cache clean --force
+#docker compose exec -T php npm install 
+#docker compose exec -d php npm run dev
+#docker compose exec -d php php artisan serve --host=0.0.0.0 --port=8000
+echo "🚀 app available at ${APP_URL}"
+
 # docker compose run node npx
